Refactor: Move coursesTreeProvider related classes to directory, renamed restclient file to restClient
