+++ vscode4teaching-extension/src/coursesTreeProvider/v4titem.ts
@@ -1,5 +1,5 @@
 import * as vscode from 'vscode';
-import { Course, Exercise } from './model/serverModel';
+import { Course, Exercise } from '../model/serverModel';
 import * as path from 'path';
 
 export class V4TItem extends vscode.TreeItem {
